Pulls blog post excerpt into fence-block
Was concerned devs were reading too much about backbone's router
and not spending their time wisely. Removed this reading as a link
and moved it to a fence block so that people know exactly what we're
asking them to do and giving them the option to read the rest should
they be interested.